feat: adding i3status to dependencies...
Adding a useless dependency to the dependencies... Why you ask? Well...
people are trying to launch i3 instead of theom and they think that
theom is broken because i3status is missing!

Not to blame the AxOS users--they are beginners. But this commit ensures
that "they wont feel that it is broken".